requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
comet_ml==3.52.1
configobj==5.0.9
contourpy==1.3.3
cycler==0.12.1
decord==0.6.0
diffusers==0.34.0
dulwich==0.24.1
einops==0.8.1
everett==3.1.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
Pygments==2.19.2
pyparsing==3.2.4
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
semantic-version==2.10.0
sentry-sdk==2.37.1
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sympy==1.14.0
tifffile==2025.9.9
tokenizers==0.21.0
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
transformers==4.53.3
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
wrapt==1.17.3
wurlitzer==3.1.1
yarl==1.20.1
zipp==3.23.0
